Reduce the size of the expected tests and move all tests relating to unmapped target policy into one test class
