--- src/test/java/org/mapstruct/intellij/inspection/UnmappedTargetPropertiesInspectionReportPolicyStaticImportTest.java
@@ -1,68 +1,0 @@
-/*
- * Copyright MapStruct Authors.
- *
- * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
- */
-package org.mapstruct.intellij.inspection;
-
-import com.intellij.codeInsight.intention.IntentionAction;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-/**
- * Tests if unmappedTargetPolicy could be read, if static import is used.
- * @author hduelme
- */
-public class UnmappedTargetPropertiesInspectionReportPolicyStaticImportTest extends BaseInspectionTest {
-
-    @NotNull
-    @Override
-    protected Class<UnmappedTargetPropertiesInspection> getInspection() {
-        return UnmappedTargetPropertiesInspection.class;
-    }
-
-    @Override
-    protected void setUp() throws Exception {
-        super.setUp();
-        myFixture.copyFileToProject(
-            "UnmappedTargetPropertiesData.java",
-            "org/example/data/UnmappedTargetPropertiesData.java"
-        );
-    }
-
-    public void testUnmappedTargetPropertiesReportPolicyStaticImport() {
-        doTest();
-        List<IntentionAction> allQuickFixes = myFixture.getAllQuickFixes();
-
-        assertThat( allQuickFixes )
-            .extracting( IntentionAction::getText )
-            .as( "Intent Text" )
-            .containsExactly(
-                "Ignore unmapped target property: 'moreTarget'",
-                "Add unmapped target property: 'moreTarget'",
-                "Ignore unmapped target property: 'testName'",
-                "Add unmapped target property: 'testName'",
-                "Ignore unmapped target property: 'testName'",
-                "Add unmapped target property: 'testName'",
-                "Ignore unmapped target property: 'moreTarget'",
-                "Add unmapped target property: 'moreTarget'",
-                "Ignore unmapped target property: 'testName'",
-                "Add unmapped target property: 'testName'",
-                "Ignore all unmapped target properties",
-                "Ignore unmapped target property: 'moreTarget'",
-                "Add unmapped target property: 'moreTarget'",
-                "Ignore unmapped target property: 'testName'",
-                "Add unmapped target property: 'testName'",
-                "Ignore all unmapped target properties",
-                "Ignore unmapped target property: 'testName'",
-                "Add unmapped target property: 'testName'",
-                "Ignore unmapped target property: 'moreTarget'",
-                "Add unmapped target property: 'moreTarget'",
-                "Ignore unmapped target property: 'moreTarget'",
-                "Add unmapped target property: 'moreTarget'"
-            );
-    }
-}
