+++ testData/inspection/UnmappedTargetPropertiesReportPolicyBeanMappingBeforeClassConfig.java
@@ -8,10 +8,6 @@ import org.example.data.UnmappedTargetPropertiesData.Source;
 import org.example.data.UnmappedTargetPropertiesData.Target;
 import org.mapstruct.*;
 
-interface NotMapStructMapper {
-
-    Target map(Source source);
-}
 
 @Mapper(unmappedTargetPolicy = ReportingPolicy.ERROR)
 interface SingleMappingMapper {
@@ -22,42 +18,13 @@ interface SingleMappingMapper {
 }
 
 @Mapper(unmappedTargetPolicy = ReportingPolicy.ERROR)
-interface SingleMappingsMapper {
-
-    @Mappings({
-        @Mapping(target = "moreTarget", source = "moreSource")
-    })
-    @BeanMapping(unmappedTargetPolicy = ReportingPolicy.WARN)
-    Target <warning descr="Unmapped target property: testName">map</warning>(Source source);
-}
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.ERROR)
-interface SingleMappingsNoBracesMapper {
-
-    @Mappings(
-        @Mapping(target = "moreTarget", source = "moreSource")
-    )
-    @BeanMapping(unmappedTargetPolicy = ReportingPolicy.WARN)
-    Target <warning descr="Unmapped target property: testName">map</warning>(Source source);
-}
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.ERROR)
 interface NoMappingMapper {
 
     @BeanMapping(unmappedTargetPolicy = ReportingPolicy.WARN)
     Target <warning descr="Unmapped target properties: moreTarget, testName">map</warning>(Source source);
 
-    @org.mapstruct.InheritInverseConfiguration
+    @InheritInverseConfiguration
     Source reverse(Target target);
-}
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.ERROR)
-interface NoMappingsMapper {
-
-    @Mappings({
-    })
-    @BeanMapping(unmappedTargetPolicy = ReportingPolicy.WARN)
-    Target <warning descr="Unmapped target properties: moreTarget, testName">map</warning>(Source source);
 }
 
 @MapperConfig(unmappedTargetPolicy = ReportingPolicy.ERROR)
@@ -83,15 +50,5 @@ interface UpdateMapper {
 interface MultiSourceUpdateMapper {
 
     @BeanMapping(unmappedTargetPolicy = ReportingPolicy.WARN)
-    void <warning descr="Unmapped target property: moreTarget">update</warning>(@MappingTarget Target moreTarget, Source source, String testName, @org.mapstruct.Context String matching);
-}
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.ERROR)
-interface SingleMappingConstantReferenceMapper {
-
-    String TEST_NAME = "testName";
-
-    @Mapping(target = TEST_NAME, source = "name")
-    @BeanMapping(unmappedTargetPolicy = ReportingPolicy.WARN)
-    Target <warning descr="Unmapped target property: moreTarget">map</warning>(Source source);
+    void <warning descr="Unmapped target property: moreTarget">update</warning>(@MappingTarget Target moreTarget, Source source, String testName, @Context String matching);
 }
