+++ testData/inspection/UnmappedTargetPropertiesReportPolicyClassLevelBeforeConfigClassWarn.java
@@ -8,11 +8,6 @@ import org.example.data.UnmappedTargetPropertiesData.Source;
 import org.example.data.UnmappedTargetPropertiesData.Target;
 import org.mapstruct.*;
 
-interface NotMapStructMapper {
-
-    Target map(Source source);
-}
-
 @Mapper(config = AllMappingsMapperConfig.class, unmappedTargetPolicy = ReportingPolicy.WARN)
 interface SingleMappingMapper {
 
@@ -21,38 +16,12 @@ interface SingleMappingMapper {
 }
 
 @Mapper(config = AllMappingsMapperConfig.class, unmappedTargetPolicy = ReportingPolicy.WARN)
-interface SingleMappingsMapper {
-
-    @Mappings({
-        @Mapping(target = "moreTarget", source = "moreSource")
-    })
-    Target <warning descr="Unmapped target property: testName">map</warning>(Source source);
-}
-
-@Mapper(config = AllMappingsMapperConfig.class, unmappedTargetPolicy = ReportingPolicy.WARN)
-interface SingleMappingsNoBracesMapper {
-
-    @Mappings(
-        @Mapping(target = "moreTarget", source = "moreSource")
-    )
-    Target <warning descr="Unmapped target property: testName">map</warning>(Source source);
-}
-
-@Mapper(config = AllMappingsMapperConfig.class, unmappedTargetPolicy = ReportingPolicy.WARN)
 interface NoMappingMapper {
 
     Target <warning descr="Unmapped target properties: moreTarget, testName">map</warning>(Source source);
 
-    @org.mapstruct.InheritInverseConfiguration
+    @InheritInverseConfiguration
     Source reverse(Target target);
-}
-
-@Mapper(config = AllMappingsMapperConfig.class, unmappedTargetPolicy = ReportingPolicy.WARN)
-interface NoMappingsMapper {
-
-    @Mappings({
-    })
-    Target <warning descr="Unmapped target properties: moreTarget, testName">map</warning>(Source source);
 }
 
 @MapperConfig(unmappedTargetPolicy = ReportingPolicy.ERROR)
@@ -75,14 +44,5 @@ interface UpdateMapper {
 @Mapper(config = AllMappingsMapperConfig.class, unmappedTargetPolicy = ReportingPolicy.WARN)
 interface MultiSourceUpdateMapper {
 
-    void <warning descr="Unmapped target property: moreTarget">update</warning>(@MappingTarget Target moreTarget, Source source, String testName, @org.mapstruct.Context String matching);
-}
-
-@Mapper(config = AllMappingsMapperConfig.class, unmappedTargetPolicy = ReportingPolicy.WARN)
-interface SingleMappingConstantReferenceMapper {
-
-    String TEST_NAME = "testName";
-
-    @Mapping(target = TEST_NAME, source = "name")
-    Target <warning descr="Unmapped target property: moreTarget">map</warning>(Source source);
+    void <warning descr="Unmapped target property: moreTarget">update</warning>(@MappingTarget Target moreTarget, Source source, String testName, @Context String matching);
 }
