+++ testData/inspection/UnmappedTargetPropertiesReportPolicyIgnoreBeanMapping.java
@@ -8,35 +8,10 @@ import org.example.data.UnmappedTargetPropertiesData.Source;
 import org.example.data.UnmappedTargetPropertiesData.Target;
 import org.mapstruct.*;
 
-interface NotMapStructMapper {
-
-    Target map(Source source);
-}
-
 @Mapper
 interface SingleMappingMapper {
 
     @Mapping(target = "testName", source = "name")
-    @BeanMapping(unmappedTargetPolicy = ReportingPolicy.IGNORE)
-    Target map(Source source);
-}
-
-@Mapper
-interface SingleMappingsMapper {
-
-    @Mappings({
-        @Mapping(target = "moreTarget", source = "moreSource")
-    })
-    @BeanMapping(unmappedTargetPolicy = ReportingPolicy.IGNORE)
-    Target map(Source source);
-}
-
-@Mapper
-interface SingleMappingsNoBracesMapper {
-
-    @Mappings(
-        @Mapping(target = "moreTarget", source = "moreSource")
-    )
     @BeanMapping(unmappedTargetPolicy = ReportingPolicy.IGNORE)
     Target map(Source source);
 }
@@ -47,17 +22,8 @@ interface NoMappingMapper {
     @BeanMapping(unmappedTargetPolicy = ReportingPolicy.IGNORE)
     Target map(Source source);
 
-    @org.mapstruct.InheritInverseConfiguration
+    @InheritInverseConfiguration
     Source reverse(Target target);
-}
-
-@Mapper
-interface NoMappingsMapper {
-
-    @Mappings({
-    })
-    @BeanMapping(unmappedTargetPolicy = ReportingPolicy.IGNORE)
-    Target map(Source source);
 }
 
 @MapperConfig
@@ -83,15 +49,5 @@ interface UpdateMapper {
 interface MultiSourceUpdateMapper {
 
     @BeanMapping(unmappedTargetPolicy = ReportingPolicy.IGNORE)
-    void update(@MappingTarget Target moreTarget, Source source, String testName, @org.mapstruct.Context String matching);
-}
-
-@Mapper
-interface SingleMappingConstantReferenceMapper {
-
-    String TEST_NAME = "testName";
-
-    @Mapping(target = TEST_NAME, source = "name")
-    @BeanMapping(unmappedTargetPolicy = ReportingPolicy.IGNORE)
-    Target map(Source source);
+    void update(@MappingTarget Target moreTarget, Source source, String testName, @Context String matching);
 }
