+++ testData/inspection/UnmappedTargetPropertiesReportPolicyIgnoreClassAnnotation.java
@@ -4,19 +4,9 @@
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-import org.mapstruct.Mapper;
-import org.mapstruct.MapperConfig;
-import org.mapstruct.Mapping;
-import org.mapstruct.MappingTarget;
-import org.mapstruct.Mappings;
-import org.mapstruct.ReportingPolicy;
 import org.example.data.UnmappedTargetPropertiesData.Target;
 import org.example.data.UnmappedTargetPropertiesData.Source;
-
-interface NotMapStructMapper {
-
-    Target map(Source source);
-}
+import org.mapstruct.*;
 
 @Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE)
 interface SingleMappingMapper {
@@ -26,38 +16,12 @@ interface SingleMappingMapper {
 }
 
 @Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE)
-interface SingleMappingsMapper {
-
-    @Mappings({
-        @Mapping(target = "moreTarget", source = "moreSource")
-    })
-    Target map(Source source);
-}
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE)
-interface SingleMappingsNoBracesMapper {
-
-    @Mappings(
-        @Mapping(target = "moreTarget", source = "moreSource")
-    )
-    Target map(Source source);
-}
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE)
 interface NoMappingMapper {
 
     Target map(Source source);
 
-    @org.mapstruct.InheritInverseConfiguration
+    @InheritInverseConfiguration
     Source reverse(Target target);
-}
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE)
-interface NoMappingsMapper {
-
-    @Mappings({
-    })
-    Target map(Source source);
 }
 
 @MapperConfig(unmappedTargetPolicy = ReportingPolicy.IGNORE)
@@ -80,14 +44,5 @@ interface UpdateMapper {
 @Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE)
 interface MultiSourceUpdateMapper {
 
-    void update(@MappingTarget Target moreTarget, Source source, String testName, @org.mapstruct.Context String matching);
-}
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE)
-interface SingleMappingConstantReferenceMapper {
-
-    String TEST_NAME = "testName";
-
-    @Mapping(target = TEST_NAME, source = "name")
-    Target map(Source source);
+    void update(@MappingTarget Target moreTarget, Source source, String testName, @Context String matching);
 }
